debian/ubuntu kernel headers: properly use available headers
We have no guarantee on what packages are or will be available on Debian
package repositories (or Ubuntu's), instead of trying to solve the
package version based on the running kernel this patch uses the meta
packages: linux-headers-cloud-amd64 for debian and linux-headers-gcp
for ubuntu.

With that the solution becomes more resilient over what is available
in the repositories always working with whatever is present.